github desktop test
in this test, i arleady have the folder. i created a git repo on it, then commit in the app, then publish directly.
conclusion: the app simply intitiate git on the project, and all you have to do are commit and publish and create a repo on github by itself.